requirements.txt
aiogram
lxml
bs4
requests
mysql-connector-python
